[Improve] Job系のサービスモジュールとTwitterAPIのモジュールを切り分け
